+++ Shapes/Shape.h
@@ -11,8 +11,8 @@ class Shape {
 public:
     virtual ~Shape() = default;
     virtual void intoPS() = 0;
-    virtual int getWidth() = 0;
-    virtual int getHeight() = 0;
+    virtual double getWidth() = 0;
+    virtual double getHeight() = 0;
 };
